+++ ixc_django_docker/settings/staging.py
@@ -2,16 +2,11 @@
 LOGGING['handlers']['logfile']['backupCount'] = 100
 
 # Enable the per-site cache.
-if django.VERSION < (1, 10):
-    MIDDLEWARE_CLASSES = \
-        ('django.middleware.cache.UpdateCacheMiddleware', ) + \
-        MIDDLEWARE_CLASSES + \
-        ('django.middleware.cache.FetchFromCacheMiddleware', )
-else:
-    MIDDLEWARE = \
-        ('django.middleware.cache.UpdateCacheMiddleware', ) + \
-        MIDDLEWARE + \
-        ('django.middleware.cache.FetchFromCacheMiddleware', )
+MIDDLEWARE = (
+    ('django.middleware.cache.UpdateCacheMiddleware', ) +
+    MIDDLEWARE +
+    ('django.middleware.cache.FetchFromCacheMiddleware', )
+)
 
 # Enable the cached template loader.
 TEMPLATES[0]['OPTIONS']['loaders'] = [